[REF] utils: dedupe get_current_branch
There were two implementations of get_current_branch: one in
utils/git.py (via subprocess) and one in utils/gh.py (via
os_exec.run + symbolic-ref), with different behaviour on a
detached HEAD.

Keep a single canonical implementation in utils/git.py, routed
through the os_exec.run wrapper. Callers use it from there
directly (utils/gh.py uses git.get_current_branch internally);
move its test alongside the other git util tests.
